feat: new Burn Facet. ERC3643OperationsFacet is empty so deleted
