+++ .changeset/maf-split-burn.md
@@ -1,0 +1,6 @@
+---
+"@hashgraph/asset-tokenization-contracts": major
+"@hashgraph/asset-tokenization-sdk": patch
+---
+
+Split a new Burn Facet. ERC3643OperationsFacet is empty so is deleted
