+++ packages/ats/contracts/Configuration.ts
@@ -88,7 +88,7 @@ export const CONTRACT_NAMES = [
   "ERC1410ManagementFacet",
   "ERC1410IssuerFacet",
   "ERC1410TokenHolderFacet",
-  "ERC1594Facet",
+  "BurnFacet",
   "DocumentationFacet",
   "ControllerFacet",
   "DiamondFacet",
@@ -125,7 +125,6 @@ export const CONTRACT_NAMES = [
   "FreezeFacet",
   "ERC3643ManagementFacet",
   "ERC3643ReadFacet",
-  "ERC3643OperationsFacet",
   "BatchControllerFacet",
   "BatchBurnFacet",
   "BatchMintFacet",
