+++ packages/ats/contracts/contracts/facets/IAsset.sol
@@ -85,8 +85,7 @@ import { IComplianceFacet } from "./compliance/IComplianceFacet.sol";
 import { IMint } from "./mint/IMint.sol";
 import { IHoldFacet } from "./hold/IHoldFacet.sol";
 import { IBatchController } from "./batchController/IBatchController.sol";
-import { IERC1594 } from "./layer_1/ERC1400/ERC1594/IERC1594.sol";
-
+import { IBurn } from "./burn/IBurn.sol";
 import { IDocumentation } from "./documentation/IDocumentation.sol";
 import { IController } from "./controller/IController.sol";
 import { IERC20Permit } from "./layer_1/ERC1400/ERC20Permit/IERC20Permit.sol";
@@ -132,6 +131,7 @@ interface IAsset is
     IERC20Votes,
     IERC1410,
     IERC3643,
+    IBurn,
     IScheduledCrossOrderedTasks,
     IScheduledSnapshots,
     IBond,
@@ -176,7 +176,6 @@ interface IAsset is
     IBatchController,
     IHoldByPartition,
     IMint,
-    IERC1594,
     IDocumentation,
     IController,
     IERC20Permit,
