+++ packages/ats/contracts/contracts/facets/burn/BurnFacet.sol
@@ -1,22 +1,22 @@
 // SPDX-License-Identifier: Apache-2.0
 pragma solidity >=0.8.0 <0.9.0;
 
-import { IERC1594 } from "./IERC1594.sol";
-import { ERC1594 } from "./ERC1594.sol";
-import { IStaticFunctionSelectors } from "../../../../infrastructure/proxy/IStaticFunctionSelectors.sol";
-import { _ERC1594_RESOLVER_KEY } from "../../../../constants/resolverKeys.sol";
+import { IBurn } from "./IBurn.sol";
+import { Burn } from "./Burn.sol";
+import { IStaticFunctionSelectors } from "../../infrastructure/proxy/IStaticFunctionSelectors.sol";
+import { _BURN_RESOLVER_KEY } from "../../constants/resolverKeys.sol";
 
 /**
- * @title ERC1594Facet
+ * @title BurnFacet
  * @author Asset Tokenization Studio Team
- * @notice Diamond facet exposing the ERC-1594 issuance and redemption surface.
- * @dev Registers three selectors: `initialize_ERC1594`, `redeem` and `redeemFrom`. Business logic
- *      lives in the `ERC1594` abstract contract.
+ * @notice Diamond facet exposing the ERC-1594 redemption and ERC-3643 burn surfaces.
+ * @dev Registers three selectors: `burn`, `redeem` and `redeemFrom`. Business logic
+ *      lives in the `Burn` abstract contract.
  */
-contract ERC1594Facet is ERC1594, IStaticFunctionSelectors {
+contract BurnFacet is Burn, IStaticFunctionSelectors {
     /// @inheritdoc IStaticFunctionSelectors
     function getStaticResolverKey() external pure override returns (bytes32 staticResolverKey_) {
-        staticResolverKey_ = _ERC1594_RESOLVER_KEY;
+        staticResolverKey_ = _BURN_RESOLVER_KEY;
     }
 
     /// @inheritdoc IStaticFunctionSelectors
@@ -26,14 +26,13 @@ contract ERC1594Facet is ERC1594, IStaticFunctionSelectors {
         unchecked {
             staticFunctionSelectors_[--selectorIndex] = this.redeemFrom.selector;
             staticFunctionSelectors_[--selectorIndex] = this.redeem.selector;
-            staticFunctionSelectors_[--selectorIndex] = this.initialize_ERC1594.selector;
+            staticFunctionSelectors_[--selectorIndex] = this.burn.selector;
         }
     }
 
     /// @inheritdoc IStaticFunctionSelectors
     function getStaticInterfaceIds() external pure override returns (bytes4[] memory staticInterfaceIds_) {
         staticInterfaceIds_ = new bytes4[](1);
-        uint256 selectorsIndex;
-        staticInterfaceIds_[selectorsIndex++] = type(IERC1594).interfaceId;
+        staticInterfaceIds_[0] = type(IBurn).interfaceId;
     }
 }
