+++ packages/ats/contracts/scripts/domain/bond/createConfiguration.ts
@@ -64,7 +64,6 @@ const BOND_FACETS = [
   "ERC1410ManagementFacet",
   "ERC1410ReadFacet",
   "ERC1410TokenHolderFacet",
-  "ERC1594Facet",
   "DocumentationFacet",
   "ControllerFacet",
   "ERC20PermitFacet",
@@ -75,10 +74,10 @@ const BOND_FACETS = [
   "BatchMintFacet",
   "BatchTransferFacet",
   "ERC3643ManagementFacet",
-  "ERC3643OperationsFacet",
   "ERC3643ReadFacet",
   "ComplianceFacet",
   "MintFacet",
+  "BurnFacet",
 
   // Clearing & Settlement
   "ClearingActionsFacet",
