+++ packages/ats/contracts/scripts/domain/bondFixedRate/createConfiguration.ts
@@ -59,7 +59,6 @@ const BOND_FIXED_RATE_FACETS = [
   "ERC1410ManagementFacet",
   "ERC1410ReadFacet",
   "ERC1410TokenHolderFacet",
-  "ERC1594Facet",
   "DocumentationFacet",
   "ControllerFacet",
   "ERC20PermitFacet",
@@ -70,10 +69,10 @@ const BOND_FIXED_RATE_FACETS = [
   "BatchMintFacet",
   "BatchTransferFacet",
   "ERC3643ManagementFacet",
-  "ERC3643OperationsFacet",
   "ERC3643ReadFacet",
   "ComplianceFacet",
   "MintFacet",
+  "BurnFacet",
 
   // Clearing & Settlement
   "ClearingActionsFacet",
