+++ packages/ats/contracts/scripts/domain/bondKpiLinkedRate/createConfiguration.ts
@@ -56,7 +56,6 @@ const BOND_KPI_LINKED_RATE_FACETS = [
   "ERC1410ManagementFacet",
   "ERC1410ReadFacet",
   "ERC1410TokenHolderFacet",
-  "ERC1594Facet",
   "DocumentationFacet",
   "ControllerFacet",
   "ERC20PermitFacet",
@@ -67,10 +66,10 @@ const BOND_KPI_LINKED_RATE_FACETS = [
   "BatchMintFacet",
   "BatchTransferFacet",
   "ERC3643ManagementFacet",
-  "ERC3643OperationsFacet",
   "ERC3643ReadFacet",
   "ComplianceFacet",
   "MintFacet",
+  "BurnFacet",
 
   // Clearing & Settlement
   "ClearingActionsFacet",
