+++ packages/ats/contracts/scripts/domain/equity/createConfiguration.ts
@@ -57,7 +57,6 @@ const EQUITY_FACETS = [
   "ERC1410ManagementFacet",
   "ERC1410ReadFacet",
   "ERC1410TokenHolderFacet",
-  "ERC1594Facet",
   "DocumentationFacet",
   "ControllerFacet",
   "ERC20PermitFacet",
@@ -68,10 +67,10 @@ const EQUITY_FACETS = [
   "BatchMintFacet",
   "BatchTransferFacet",
   "ERC3643ManagementFacet",
-  "ERC3643OperationsFacet",
   "ERC3643ReadFacet",
   "ComplianceFacet",
   "MintFacet",
+  "BurnFacet",
 
   // Clearing & Settlement (8)
   "ClearingActionsFacet",
