+++ packages/ats/contracts/scripts/domain/loan/createConfiguration.ts
@@ -63,7 +63,6 @@ const LOAN_FACETS = [
   "CoreAdjustedFacet",
   "ERC20PermitFacet",
   "ERC20VotesFacet",
-  "ERC1594Facet",
   "DocumentationFacet",
   "ControllerFacet",
   "ERC1410ReadFacet",
@@ -71,7 +70,6 @@ const LOAN_FACETS = [
   "ERC1410IssuerFacet",
   "ERC1410TokenHolderFacet",
   "ERC3643ManagementFacet",
-  "ERC3643OperationsFacet",
   "ERC3643ReadFacet",
   "BatchControllerFacet",
   "BatchBurnFacet",
@@ -79,6 +77,7 @@ const LOAN_FACETS = [
   "BatchTransferFacet",
   "ComplianceFacet",
   "MintFacet",
+  "BurnFacet",
 
   // Hold
   "HoldFacet",
