+++ packages/ats/contracts/scripts/domain/orchestratorLibraries.ts
@@ -109,12 +109,12 @@ export const LIBRARY_DEPENDENT_FACETS: Record<string, Array<keyof typeof LIBRARY
   ERC1410TokenHolderFacet: ["tokenCoreOps"],
   ERC1410ReadFacet: ["tokenCoreOps"],
   ERC1410IssuerFacet: ["tokenCoreOps"],
-  ERC1594Facet: ["tokenCoreOps"],
   ControllerFacet: ["tokenCoreOps"],
   BatchControllerFacet: ["tokenCoreOps"],
   BatchBurnFacet: ["tokenCoreOps"],
   BatchTransferFacet: ["tokenCoreOps"],
-  ERC3643OperationsFacet: ["tokenCoreOps"],
+  MintFacet: ["tokenCoreOps"],
+  BurnFacet: ["tokenCoreOps"],
   // HoldOps dependencies - hold/lock operations
   HoldManagementFacet: ["holdOps"],
   HoldFacet: ["holdOps"],
