+++ packages/ats/contracts/test/fixtures/tokens/loansPortfolio.fixture.ts
@@ -31,7 +31,7 @@ import {
   ERC3643ManagementFacet__factory,
   ERC20VotesFacet__factory,
   ControllerFacet__factory,
-  ERC1594Facet__factory,
+  MintFacet__factory,
   ERC1410ManagementFacet__factory,
   FreezeFacet__factory,
   KycFacet__factory,
@@ -137,7 +137,7 @@ export async function deployLoansPortfolioTokenFixture({
   const freezeFacet = FreezeFacet__factory.connect(proxyAddress, deployer);
   const capFacet = CapFacet__factory.connect(proxyAddress, deployer);
   const controllerFacet = ControllerFacet__factory.connect(proxyAddress, deployer);
-  const erc1594Facet = ERC1594Facet__factory.connect(proxyAddress, deployer);
+  const mintFacet = MintFacet__factory.connect(proxyAddress, deployer);
   const erc1410ManagementFacet = ERC1410ManagementFacet__factory.connect(proxyAddress, deployer);
   const erc3643ManagementFacet = ERC3643ManagementFacet__factory.connect(proxyAddress, deployer);
   const erc20VotesFacet = ERC20VotesFacet__factory.connect(proxyAddress, deployer);
@@ -165,7 +165,7 @@ export async function deployLoansPortfolioTokenFixture({
     },
     securityType: 1, // SecurityType.Equity (reuse for loan portfolio)
   });
-  await erc1594Facet.initialize_ERC1594();
+  await mintFacet.initialize_ERC1594();
   await capFacet.initialize_Cap(securityData.maxSupply, []);
   await protectedPartitionsFacet.initialize_ProtectedPartitions(securityData.arePartitionsProtected);
   await clearingActionsFacet.initializeClearing(securityData.clearingActive);
